adjusted basic Middlewares
